backend/requirements.txt
fastapi
uvicorn
vllm
pydantic
python-dotenv
transformers
jinja2